body
program is called 100-lockers-puzzle. 
it is a representation of the 100 lockers puzzle.
program takes from the user the number of doors that are to be filled into the hallway. lets say we want to have a puzzle with 500 doors.
displays an array of open or closed doors represented by 1 and 0 respectively.
displays number of open and closed doors, ratio of open doors to closed doors and prompts the user for a door number and returns the state of that door.